+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Possibility to create model in the middle of the sequence
+
 ## [0.4.1] - 2023-02-08
 
 ### Added