+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rename method `getInitPosition` to `startPosition`
+- Rename method `move` to `shift`
 - Models now are shifted after the model update
 
 ## [0.4.1] - 2023-02-08