+++ CHANGELOG.md
@@ -6,19 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.5.0] - 2023-02-09
 
 ### Added
 
 - Possibility to create model in the middle of the sequence
 - Possibility to create model in the beginning of the sequence
-- Extra argument for shift amount in `shiftToStart` and `shiftToEnd` methods  
 - Possibility to update positions without shifting other models
+- Extra argument for shift amount in `shiftToStart` and `shiftToEnd` methods  
 
 ### Changed
 
 - Rename method `getInitPosition` to `startPosition`
-- Models now are shifted after the model update
+- Position of other models now are shifted after the model update
 
 ## [0.4.1] - 2023-02-08